services/vanna/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
vanna[postgres]==0.5.0
groq==0.4.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==1.10.13
